MissionSrijan26_FINAL (1)/srijan_final/requirements.txt
streamlit==1.32.0
langchain==0.1.20
langchain-groq==0.1.3
groq==0.5.0
requests==2.31.0
playwright==1.42.0
python-dotenv==1.0.1
